+++ server/src/lib.rs
@@ -1,4 +1,5 @@
-use sqlx::PgPool;
+use ledger::Ledger;
+use sqlx::{PgPool, postgres::PgPoolOptions};
 use std::sync::Arc;
 
 pub mod constants;
@@ -45,4 +46,36 @@ pub struct Config {
 pub enum SubCommand {
     /// Starts the server
     Start,
+}
+
+impl ApiContext {
+    pub async fn new(config: Config) -> anyhow::Result<ApiContext> {
+        let mainnet_ledger = Ledger::init_async(&config.mainnet_ledger_path)
+        .await
+        .expect("Error: Could not create mainnet ledger.");
+
+        let devnet_ledger = Ledger::init_async(&config.devnet_ledger_path)
+        .await
+        .expect("Error: Could not create devnet ledger.");
+
+        let mainnet_db = PgPoolOptions::new()
+        .max_connections(25)
+        .connect(&config.mainnet_database_url)
+        .await
+        .context("Error: Could not connect to mainnet database.")?;
+
+        let devnet_db = PgPoolOptions::new()
+        .max_connections(25)
+        .connect(&config.devnet_database_url)
+        .await
+        .context("Error: Could not connect to devnet database.")?;
+
+        Ok(ApiContext {
+            config: Arc::new(config),
+            mainnet_ledger: Arc::new(mainnet_ledger.db),
+            devnet_ledger: Arc::new(devnet_ledger.db),
+            mainnet_db,
+            devnet_db,
+        })
+    }
 }
